Updating the name of the generated kernel module
This commit updates the name of the km generated as a result of
the nexmon build process. This change makes it easy to deploy a
patched version of the brcmfmac driver w/o affecting the original
version or causing conflicts b/w debain pkgs that provide the
original and the patched version of the driver.